+++ src/CHTL/CHTLNode/BaseNode.h
@@ -300,6 +300,9 @@ private:
     String alias_;
 };
 
+// 前向声明
+class ConditionalNode;
+
 // 访问者接口
 class NodeVisitor {
 public:
@@ -316,6 +319,7 @@ public:
     virtual void visit(CustomNode& node) = 0;
     virtual void visit(OriginNode& node) = 0;
     virtual void visit(ImportNode& node) = 0;
+    virtual void visit(ConditionalNode& node) { (void)node; } // 默认实现，子类可选重写
 };
 
 } // namespace CHTL